refactored main file, reduced from ~2000 lines to 679, added lottie load animation
